Make wslay compilable for MSVC

+++ lib/wslay_event.c
@@ -629,7 +629,7 @@ int wslay_event_recv(wslay_event_context_ptr ctx)
         if(!wslay_event_config_get_no_buffering(ctx) ||
            wslay_is_ctrl_frame(iocb.opcode)) {
           if((r = wslay_event_imsg_append_chunk(ctx->imsg,
-                                                iocb.payload_length)) != 0) {
+                                                (size_t) iocb.payload_length)) != 0) {
             ctx->read_enabled = 0;
             return r;
           }
@@ -666,7 +666,7 @@ int wslay_event_recv(wslay_event_context_ptr ctx)
            wslay_is_ctrl_frame(iocb.opcode)) {
           struct wslay_event_byte_chunk *chunk;
           chunk = wslay_queue_tail(ctx->imsg->chunks);
-          wslay_event_byte_chunk_copy(chunk, ctx->ipayloadoff,
+          wslay_event_byte_chunk_copy(chunk, (size_t) ctx->ipayloadoff,
                                       iocb.data, iocb.data_length);
         }
         ctx->ipayloadoff += iocb.data_length;
@@ -847,7 +847,7 @@ int wslay_event_send(wslay_event_context_ptr ctx)
       iocb.rsv = ctx->omsg->rsv;
       iocb.mask = ctx->server^1;
       iocb.data = ctx->omsg->data+ctx->opayloadoff;
-      iocb.data_length = ctx->opayloadlen-ctx->opayloadoff;
+      iocb.data_length = (size_t) (ctx->opayloadlen-ctx->opayloadoff);
       iocb.payload_length = ctx->opayloadlen;
       r = wslay_frame_send(ctx->frame_ctx, &iocb);
       if(r >= 0) {
